Move DatadogReporterFactory to correct package
