+++ dropwizard-metrics-datadog/src/main/java/io/dropwizard/metrics/common/DatadogReporterFactory.java
@@ -1,10 +1,9 @@
-package io.dropwizard.metrics;
+package io.dropwizard.metrics.common;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.ScheduledReporter;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.dropwizard.metrics.common.BaseReporterFactory;
 import org.staircase.metrics.datadog.DatadogReporter;
 import org.staircase.metrics.datadog.DefaultMetricNameFormatterFactory;
 import org.staircase.metrics.datadog.DynamicTagsCallbackFactory;
