--- dropwizard-metrics-datadog/src/main/resources/META-INF/services/io.dropwizard.metrics.ReporterFactory
@@ -1,1 +1,0 @@
-io.dropwizard.metrics.DatadogReporterFactory
