+++ dropwizard-metrics-datadog/src/main/resources/META-INF/services/io.dropwizard.metrics.common.ReporterFactory
@@ -1,0 +1,1 @@
+io.dropwizard.metrics.common.DatadogReporterFactory
